Move logic to request ice servers into factory class
